requirements.txt
langchain
openai
python-dotenv
chromadb
tiktoken